Add 'srcdoc' check for URI, allowing HTML to be passed in
Currently the package allows a developer to pass in a url as a 'src'
which then gets parsed and passed into the WebViewController's
.loadRequest method. This PR adds a check that looks for 'srcdoc' to
the flow. The change would allow for html to be passed into the iframe
under the 'srcdoc' property. Standard practice is to check for 'srcdoc'
first and if one is not present we fallback to 'src'.

This change requires the srcUri to be constructed outside of the
returning the SizedBox widget.

This change would allow a developer to pass custom html into the
iframe. One scenario for this could be that the html needs to be
modified before it is displayed by adjusting the content they want to
display or adding some Javascript logic.

For current developers, the usage is the same so this PR is not a
breaking changes, but it will add some an additional feature that could
prove useful to some.