Allow height and width assignment to iframe
This PR adds a small helper class called IframeProperties that
currently includes height and width properties. It also updates the
iframe widgets to take an IframeProperties parameter. This will allow
developers to control the size of the SizedBox which is the parent of
the WebViewWidget created to display the url/html for the iframe.

In IframeWidget, when setting the givenWidth and givenHeight properties
we first check if the iframeProperties width and height are not null,
if they have a value we use that, otherwise we continue with the logic
from before.